moved delete content to data section
